basics: Add single data file with a Martini recipe
Tried to work out the semantics of a very simple recipe, albeit one with
multiple steps, pieces of required equipment, ingredients, units, etc.

It's not simple! And I can see a lot of places where it could get more
complex. Probably the single biggest thing I did was to make
"ingredients" and "equipment" basically the same thing. Not sure if this
is the greatest idea in the long run but it did simplify it in my head.